generate_pet_name/requirements.txt
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
langchain-core==0.3.75
langchain-google-genai==2.1.10
python-dotenv==1.1.1
streamlit==1.49.1

